ft(skills section):
changes all the files into the jsx extension
add a favicon

[ Maintains #172320184 ]
