+++ package.json
@@ -47,6 +47,7 @@
     "prettier": "^2.0.4",
     "sass-loader": "^8.0.2",
     "style-loader": "^1.1.4",
+    "url-loader": "^4.1.0",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3"
