+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Navbar from './components/navbar/Navbar'
-import MainContent from './components/mainContent'
+import Navbar from './components/navbar/Navbar.jsx'
+import MainContent from './components/mainContent.jsx'
 import './components/styles/main.scss'
 
 const App = () => {
