+++ src/components/styles/about.scss
@@ -14,6 +14,7 @@
     @include flexCenter(column, null, null);
     padding: 5rem 19rem;
     text-align: center;
+    width: auto;
   
     #{&}_heading {
       font-size: 34px;
